+++ computing-in-context-frontend/package.json
@@ -8,7 +8,7 @@
     "start": "next start",
     "lint": "next lint",
     "upload": "npx tsx src/scripts/uploadResourcesToDrive.ts",
-    "embed": "npx tsx src/scripts/saveResources.ts"
+    "embed": "npx tsx src/scripts/embedResources.ts"
   },
   "dependencies": {
     "@nteract/commutable": "^7.5.1",
