+++ computing-in-context-frontend/src/scripts/embedResources.ts
